fix: only allow editing publish_as if you have the current mandate
I think it would make sense to be able to change publish_as in other cases too, but
that would require fetching their names from dfunkt separately, so this seems like a
least-effort good-enough solution

With the behaviour before this commit, you could not change the mandate of an existing
post, even if you just created it (which to me seems unneccessarily restrictive). But
it seemed like you could in the frontend.